Moving NgTestListener
